refactor: modify error message in toast, and move edit hook
